Adjust weapon rarity system and refactor related UI elements
The main adjustment is in the weapon rarity system where the ranking of the rarity is now part of the upgrade name. The weapon panel background color now also changes according to the weapon's rarity.

For the weapon upgrades and items, the application of rarity was changed. Additionally, the upgrades and items now use "DamageIncreasePercentage" instead of static addition for the description text and calculation.

A rounding issue for tornado, lightning strike, and book projectile was addressed where the Damage value is now a float instead of an int to conform with the original weapon damage type.

For the upgrade panel, there's an increase in the random range for upgrade entries display. In the cases of Black Hole, Fireball, Dark Orb, and Tornado upgrades, the description of the upgrade now shows percentage increase instead of static addition.

EnemyManager's damage no longer uses int but rather float to encourage damage variations among weapons.

The size and position of some components in Essential.unity were adjusted for better user interface.

The player's health regen was changed to a smaller increment per second resulting in a smoother health increase.

LightningStrikeProjectile's anchored position is adjusted and SetActvie(false) codes from upgrade entries are removed to prevent disappearance from list.

The changes enhance game dynamics and improve user interface.