Add dodge chance stat to player characters
This commit introduces a new "dodge" stat to the player characters. This stat gives each character a certain chance to dodge incoming damage. It can be increased by items that the character obtains during the game. A new stat factoring logic is also implemented to prevent the armor increase when there is no basic armor stat. The CharacterInfoPanel now uses the PlayerStats class for retrieving the stats list instead of characterData.GetStatsList(), reflecting the transition of stat handling to the PlayerStats class. This change was made to support the new complexity introduced by the dodge stat.

Please note that some characters will have special effects associated with this dodge stat. For example, characters of rank E1 or higher have a 0.1 increase in dodge chance.