Add ChitoseStrategy, ICharacterStrategy, and PlayerStatsUpdater meta files
Created meta files for ChitoseStrategy, ICharacterStrategy, and PlayerStatsUpdater. These files are needed for establishing a proper relationship with Unity engine and managing the serialization of the associated assets. The meta files consist of the unique guid to identify the script files, version of file format and other required default configurations.