Add DiscordGameSDK for game management.
This large commit introduces the DiscordGameSDK to the codebase to improve integration with Discord. It includes classes, methods, and comprehensive structures that the SDK provides for improving social features. This allows interaction with Discord in various ways, such as updating activities, managing storage and stores, joining and creating lobbies, etc. The SDK can help with managing Discord related features in a unified way.
Also added metadata files associated with the new files for Unity’s internal use.