Add toggleable filtering to weapons and items management
This commit introduces `WeaponToggleableEntry` and `ItemToggleableEntry` classes. They extend the new `ToggleableEntry` base class and allow the game to toggle whether specific Weapons and Items are currently enabled for player use.

The commit updates the `WeaponManager` and `PlayerStatsComponent` classes, replacing the lists of `WeaponBase` and `ItemBase` directly utilized with the new toggleable entry subclasses. Changes in the `WeaponManager` ensure the toggled state of items or weapons is respected when adding, removing, or determining available upgrades.

These changes have been implemented to allow selective enabling/disabling of game content. This fulfills a part of the planned features for the game's content management system. Alongside this, irrelevant variables and methods were removed from `CharacterInfoPanel.cs`, `WeaponManagerEditor.cs` was added for better content management in Unity Editor, and minor optimizations were made in `ChitoseStrategy.cs` and `UpgradeEntry.cs`. `ReadmeEditor.cs` was fixed replacing `Editor` class reference with `UnityEditor.Editor`.

To reflect the changes in code logic, the related UI components are updated. Specifically, the 'Character Info Panel' prefab and the 'Main Menu' scene were modified to work with the updated `CharacterInfoPanel.cs` behaviour.