Add 'SpecialBar' for controlling unique character features
A new 'SpecialBar' feature is added along with related elements in the UI and GameManager. It's a unique resource bar for characters with special abilities, giving an additional level of gameplay complexity. The 'SpecialBar' can be adjusted and checked for max value. This feature is linked to the save file and various gameplay objects such as enemies. Relevant graphics files have also been updated to visually represent the SpecialBar in the game.