Add Amelia's skill and damage taken event
Introduce Amelia's unique skill, AmeliaGlassShield, which is created on PlayerSkillComponent's startup if the character played is Amelia. Add corresponding AmeliaGlassShard for the shield. Add DamageTakenStaticEvent, allowing other classes to subscribe to this event to implement various reactions when the damage is taken. It is now used in AmeliaGlassShield to destroy a shard when the player character takes damage, and heals equivalent amount of health. Apply the changes to Amelia's character data and also adjust color in HealingLightBeam material.