Add events and logic for Special Bar and Experience pickup
Added a new SpecialBarFilledEvent script to handle events when the Special Bar is filled, along with an ISpecialBarFilledHandler interface. Made corresponding changes to the SpecialBarManager script.

Implemented Increment functions in the SpecialBarManager which increases the bar's value. Update function modified to trigger SpecialBarFilledEvent when the bar is full.

Renamed DamageTakenStaticEvent script to DamageTakenEvent. Corresponding changes made in other scripts and Unity files.

On picking up Experience, the ExpPickedUpEvent is invoked. Also, increment the Special Bar when experience is picked up.

AmeliaGlassShield script updated to implement these events. Now shards are created on picking up experience or when the special bar is filled, and not just on Awake. Also made some minor adjustments to Amelia's Special Bar values and shard spawning parameters.

Apply these changes to improve game mechanisms, allowing a new layer of strategies and interaction through the use of the Special Bar and Experience.