Add Story Entry viewing and new Blur Manager
StoryTile.cs now maintains a serialized private LoreEntry and a new func OpenStoryEntry(). LoreEntryPanel.cs, LoreEntryContainer and CharacterLoreEntry are updated to work with new LoreEntry. CursorManager calls are removed in UI managements. A new BlurManager is added to handle Menu and Game blur functionalities. Corresponding changes are made in prefabs and Unity scenes for the new logic.