Optimize NishiSkill damage calculation
Moved the calculation for NishiSkill's damage from the Update method to the Start method. The damage value is now stored in a private variable, which is used when enemies take damage. This reduces the number of times this calculation is done from every frame to only once at the start, optimising game performance without altering functionality.