Add unlock conditions to story entries
Added a feature to display lock icons for story sections that require a certain number of story points to unlock. Story points are now gathered based on the amount of time played in game. Also made some minor adjustments to the UI layout to include the player's progress towards unlocking new story sections.

This feature was implemented in order to provide players with a sense of progression and reward for spending time in-game, as well as to create an element of mystery around locked story sections, thus encouraging continued playing.