Update weapon and item handling with containers
Refactored how available weapons and items are stored in WeaponManager.cs. Removed serialized lists and replaced them with WeaponContainer and ItemContainer objects. Now, weapon and item containers are responsible for the initialization of the availableWeapons and availableItems lists. This change provides better encapsulation and makes it more straightforward to add or remove weapons/items.

Added ContainerUtility.cs, which enables enable/disable all actions for the item and weapon lists in Unity Editor. It improves ease of work within Unity Editor by providing quick enable/disable of all items in a container.

Added IsUnlocked method to IPlayerItem Interface which checks if the items are unlocked in the provided save file.

Edited Main Menu screen within Unity Editor and added scripts related to handling the ContainerUtility in the ItemContainerEditor.cs and updated unlockables screen with additional features.

Applied changes in MainMenu.unity scene to reflect new logic.

All changes encourage better organization and easier manipulation of weapons and items within the Unity Editor and gameplay.