Remove Gacha Image.prefab, add DragAnimator.cs, update Sprite Shadow shadergraph
The 'Gacha Image.prefab' was deleted because it was not used in this project anymore. A new script 'DragAnimator.cs' has been added to handle animation of the UI elements using drag functionality. 'Sprite Shadow.shadergraph' has been updated due to changes in the actual shader's logic which include a new method to manipulate shadows on the character sprite. This update will enhance shadow effects in the game interface.