Refactor code, move CustomizeAvailableItemList method into relevant editors
The aim of this commit was to delete the `ContainerUtility.cs` file and move its `CustomizeAvailableItemList` method directly into the `ItemContainerEditor.cs` and `WeaponContainerEditor.cs` files. This ensures that these editor files are more inclusive and independence, removing the necessity for an extra script dedicated to a utility function. Moreover, the `preloadedAssets` array in the `ProjectSettings.asset` was modified to be empty, removing outdated IDs.