Add new materials for Arrow Rain and Bouncer abilities
This commit introduces materials for Arrow Rain and Bouncer abilities, along with their respective prefabs. These materials are used to improve the visual representation of the abilities in the game. The shaders for the sprite shadow have also been updated to match with the new materials. Some unnecessary blocks and slots in the shaders were removed as part of the cleanup.