Remove Magicball icon and add new ability assets
This commit includes the removal of the Magicball icon from the project assets due to its redundancy.

We've also introduced new icon assets for abilities Arrow Rain, Ground Slash, Bouncer, and SpaceExpansionBall, along with updating their respective prefab files to reference these new icons. New Wind Shear ability assets were added to the project with the associated attributes set in the Material.mat file.

The scale down effect used in the SpaceBallProjectile.cs file was adjusted to create a smoother animation. The adjustment was made by reducing the size of the projectile at a faster rate and decreasing the wait time between reductions.

These changes are implemented to improve visual representation and user experience in the game.