Remove Amelia avatar and update various game settings and assets
This commit involves several changes aimed at improving our game performance and player experience. Amelia's avatar has been deleted as it is not part of the game anymore. The color of the Sprite Outline Material was changed from a dark blue to black, providing better visual contrast in-game. With respect to the Ground Slash Weapon prefab, the AttackCount has been increased to 1 for better damage dealing, and the availableUpgrades property was populated with actual upgrades.

Also, there was a change in the gems, food, and gold pickups prefabs, where the associated material fileID was updated, and trigger-related settings were altered. In the Damage Number prefab, text size was increased for better readability, and the FireBallWeapon.cs script was updated to disable the audio upon the fireball's spawn - we'll be using a different audio feedback mechanism.

Additionally, this commit introduces new hit sounds, and changes the ThrowingKnifeProjectile.cs to use simple damage. Last but not least, changes in AmeliaExplosion.cs making the explosion dependent on the elapsed time rather than a remaining time count down.

All these modifications aim to enhance the game's visual quality, sound effects, and interaction, making our game more engaging and enjoyable for our players.