Remove checkbox.png.meta and update AchievementEnum and various other files
The file checkbox.png.meta was removed due to it being obsolete and no longer used within our project. AchievementEnum.cs was updated to refine the descriptions of the achievements for better player understanding and immersion. Other miscellaneous changes include the addition of a mask to the Upgrade Entry Panel for better UI visualization, and minor alterations to button graphics and game object attributes for improved consistency and performance.