Add floating, animated shard displays in the recollection panel
This commit improves the display of shard characters in the recollection panel of the main menu. Instead of using three separate CharacterDisplayPanels, it now uses a list of CharacterDisplayPanels to enable dynamic display of 5 shard characters. The CharacterDisplayPanel also has been enhanced to include a floating animation. The floating Y position is calculated using a sine function to create a smooth up and down movement. Also, gem text display in the Recollection Menu is updated to highlight the gems number with a specific color. New images/meta files are also added for these new improvements. These changes enhance the user's visual experience, enabling a broader view of available characters while also making the display more dynamic and visually appealing.