Refactor laser gun and enemy mechanics
This commit revises the methods of the LaserGunProjectile to improve performance and readability. The LaserGunProjectile class is updated to interact with the new method in the Enemy class, ensuring it acts only on enemies that are not dying. In LaserGunWeapon, the SetupProjectile method is implemented replacing the previous ProjectileSpawn. The LaserGunNetworkComponent changes include methods to adequately deal with line rendering.